+++ lib/healthcheck/worker_experimental_test.go
@@ -1,5 +1,3 @@
-//go:build go1.24 && enablesynctest
-
 /*
  * Teleport
  * Copyright (C) 2025  Gravitational, Inc.
@@ -70,12 +68,8 @@ func TestGetTargetHealth(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.desc, func(t *testing.T) {
-			synctest.Run(func() {
-				// TODO(gavin): use t.Context() once we bump to go1.25 which adds synctest.Test(t, func(t *testing.T) { ... })
-				// See https://github.com/golang/go/issues/74837#issuecomment-3177146173 for more info
-				ctx, cancel := context.WithCancel(context.Background())
-				defer cancel()
-				worker, err := newWorker(ctx, workerConfig{
+			synctest.Test(t, func(t *testing.T) {
+				worker, err := newWorker(t.Context(), workerConfig{
 					HealthCheckCfg: test.healthCheckConfig,
 					Target: Target{
 						GetResource: func() types.ResourceWithLabels { return nil },
