+++ lib/integrations/awsra/profile_syncer_test.go
@@ -1,5 +1,3 @@
-//go:build go1.24 && enablesynctest
-
 /*
  * Teleport
  * Copyright (C) 2025  Gravitational, Inc.
@@ -244,18 +242,16 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 			tags: exampleProfileTags,
 		}
 
-		synctest.Run(func() {
-			ctx, cancel := context.WithCancel(context.Background())
+		synctest.Test(t, func(t *testing.T) {
 			go func() {
-				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(ctx, params)
+				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(t.Context(), params)
 				assert.NoError(t, err)
 			}()
 
 			// Wait for the 1st profile sync iteration.
 			synctest.Wait()
-			cancel()
 
-			require.Len(t, serverClient.appServers, 0)
+			require.Empty(t, serverClient.appServers)
 		})
 	})
 
@@ -272,16 +268,14 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 			tags: exampleProfileTags,
 		}
 
-		synctest.Run(func() {
-			ctx, cancel := context.WithCancel(context.Background())
+		synctest.Test(t, func(t *testing.T) {
 			go func() {
-				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(ctx, params)
+				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(t.Context(), params)
 				assert.NoError(t, err)
 			}()
 
 			// Wait for the 1st profile sync iteration.
 			synctest.Wait()
-			cancel()
 
 			require.Len(t, serverClient.appServers, 1)
 			appServer := serverClient.appServers[0]
@@ -297,16 +291,14 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 				"teleport.dev/integration":                    "test-integration",
 			}, appServer.GetAllLabels())
 
-			t.Run("integration status is updated", func(t *testing.T) {
-				status := serverClient.integrations[integrationWithProfileSync.GetName()].GetStatus()
-				require.NotNil(t, status)
-				lastSyncSummary := status.AWSRolesAnywhere.LastProfileSync
-				require.Equal(t, types.IntegrationAWSRolesAnywhereProfileSyncStatusSuccess, lastSyncSummary.Status)
-				require.NotEmpty(t, lastSyncSummary.StartTime)
-				require.NotEmpty(t, lastSyncSummary.EndTime)
-				require.Equal(t, int32(1), lastSyncSummary.SyncedProfiles)
-				require.Empty(t, lastSyncSummary.ErrorMessage)
-			})
+			status := serverClient.integrations[integrationWithProfileSync.GetName()].GetStatus()
+			require.NotNil(t, status)
+			lastSyncSummary := status.AWSRolesAnywhere.LastProfileSync
+			require.Equal(t, types.IntegrationAWSRolesAnywhereProfileSyncStatusSuccess, lastSyncSummary.Status)
+			require.NotEmpty(t, lastSyncSummary.StartTime)
+			require.NotEmpty(t, lastSyncSummary.EndTime)
+			require.Equal(t, int32(1), lastSyncSummary.SyncedProfiles)
+			require.Empty(t, lastSyncSummary.ErrorMessage)
 		})
 	})
 
@@ -326,16 +318,14 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 			tags: tags,
 		}
 
-		synctest.Run(func() {
-			ctx, cancel := context.WithCancel(context.Background())
+		synctest.Test(t, func(t *testing.T) {
 			go func() {
-				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(ctx, params)
+				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(t.Context(), params)
 				assert.NoError(t, err)
 			}()
 
 			// Wait for the 1st profile sync iteration.
 			synctest.Wait()
-			cancel()
 
 			require.Len(t, serverClient.appServers, 1)
 			appServer := serverClient.appServers[0]
@@ -369,16 +359,14 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 			tags: tags,
 		}
 
-		synctest.Run(func() {
-			ctx, cancel := context.WithCancel(context.Background())
+		synctest.Test(t, func(t *testing.T) {
 			go func() {
-				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(ctx, params)
+				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(t.Context(), params)
 				assert.NoError(t, err)
 			}()
 
 			// Wait for the 1st profile sync iteration.
 			synctest.Wait()
-			cancel()
 
 			status := serverClient.integrations[integrationWithProfileSync.GetName()].GetStatus()
 			require.NotNil(t, status)
