+++ lib/srv/discovery/discovery_eks_test.go
@@ -1,5 +1,3 @@
-//go:build go1.24 && enablesynctest
-
 /*
  * Teleport
  * Copyright (C) 2023  Gravitational, Inc.
@@ -248,9 +246,8 @@ func TestDiscoveryServerEKS(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			ctx, cancel := context.WithCancel(t.Context())
-
-			synctest.Run(func() {
+			synctest.Test(t, func(t *testing.T) {
+				ctx := t.Context()
 				fakeConfigProvider := mocks.AWSConfigProvider{
 					OIDCIntegrationClient: &mocks.FakeOIDCIntegrationClient{
 						Integration: awsOIDCIntegration,
@@ -292,11 +289,8 @@ func TestDiscoveryServerEKS(t *testing.T) {
 				// Wait for the discovery server to complete one iteration of discovering resources
 				synctest.Wait()
 
-				// Start server shutdown.
-				cancel()
-
 				// Discovery usage events are reported.
-				require.Greater(t, len(mockAccessPoint.usageEvents), 0)
+				require.NotEmpty(t, mockAccessPoint.usageEvents)
 
 				// Check the UserTasks created by the discovery server.
 				existingTasks := slices.Collect(maps.Values(mockAccessPoint.storeUserTasks))
